Rework autodetection to remove possible ambiguity
(Previous version relied on memory address 0xF003 on W5200 being
non-zero. It seemed to invariably be 0x63 on my device but that's a
bit of an assumption to rely on. Now we check that the W5100 mode
register really does behave exactly like a mode register, before we
assume we have a W5100.)